Mannequin Detail Tweak
Makes it so that when you use a item on a mannequin you see in the text that your trying to put pants on the mannequins head.

Update mannequin.dm

Update mannequin.dm

Unequpping the mannequin is broken except for when im swapping clothing.

Update mannequin.dm

more fix

Update mannequin.dm

New issue. Placing a shirt on the stomch will cause it not to go on the mannequin shirt slot but create a new stomach slot.

Update mannequin.dm

Makes mannequip equip unequip if existing item

Update mannequin.dm

Attempt to fix mannequin unequip for Topic proc not working. Equipping the stuff works but swapping a item with another causes the item to remain in the contents of the mannequin.