qatil merc class
The Qatil are assassins that add to the merc roster by fulfilling the previously empty role of speedy, knife wielding assassins.

They're moreso suited to direct killing as they are slightly less good at being stealthy compared to something like a rogue or antag assassin, but are able to fulfill the role of stealth regardless.