Mount Decapitation Overhaul (You should FEAR the mountain) (#942)
* Part 1 - Lava, Scarcity, and Mason Camp

Added the lava back instead of water. I'll probably add spots with murk to compensate for the lack of water. I've also made the southern resource abundancy completely gone to incentivize pushing further outwards into danger in order to find better ore, as well as moved the mason camp towards the entrance. Makes more sense that way- i've left the old camp as ruins for people to just poke their heads into as well as fight trolls at (the goal is to be more dangerous, after all.)

* Update smalldecap.dmm

* Necromancer's Den

Adds a necromancers den with alchemical ingredients and a few interesting items. These skeletons are not pushovers, however. Tread carefully.

* Beespider Cavern

Adds to the southern part of the map near the troll spot a little cavern with a barricaded entrance. Has a few beespiders to attack. A great spot for silk and a few little nick-nacks to find.

* Destroyed Grenzelhoft Camp

An outpost made by what appears to be a scouting party of the Black Empire. Why exactly they were there, who knows, but it seems that Graggar's spawn have destroyed it. The loot is VERY good if you can get it, but good luck. You're going to need friends for this.

* Update smalldecap.dmm

* Necromancer's Lair

Probably the biggest and final real addition to decapitation is the lair. It has some of the most richest materials as well as very, very good loot, but it is also ridiculously hard. Gather your drifter friends and go forth, and good luck.

* Update smalldecap.dmm

* Update smalldecap.dmm

* Compromise Pt1

This removes the Grenzel camp due to maintainer concerns and also lightens the loot load a little at my own personal decision. Also removes one of the trolls at the abandoned camp.

* Compromise pt2

Lightens mob count, and lowers the loot table more.

* danger increase to main dungeon, loot decrease to main dungeon

just removes one of the gold pile and bumps an extra 2 skeletons to the most valuable rooms. after this first round of it we'll see how it goes and whether i need to make more tweaks.

* Loot Table Nerf

Strengthens the mobs and lowers the loot table value more after some playtesting.

* The Trollening

Churn solos this. Countering it with less skeletons. More trolls-- those gave the biggest issue to adventurers.

* The Defortifying

Removes a lot of the iron bars sectioning rooms in order to make it so the mobs can and will mob you if you are alone.

* Update smalldecap.dmm

* Update smalldecap.dmm

* Update smalldecap.dmm

* Update smalldecap.dmm

* Update smalldecap.dmm

* Update smalldecap.dmm

* Update smalldecap.dmm

* Increases difficult a little of big dungeon. More mantraps too because im a prick.