Tailor Shop Changes (And minor merchant changes) (#722)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Somewhat Re-does the layout of the Tailor shop, making it function
better.
The Merchant's stock of fancy clothing has gone missing! The thief has
left behind a few armor stands.

Tailor Shop

![image](https://github.com/user-attachments/assets/b2d9f819-d756-45c1-ae7b-32d5955d405e)

![image](https://github.com/user-attachments/assets/31755822-7f83-486f-bc37-28b91b229e50)
Merchant Shop

![image](https://github.com/user-attachments/assets/6501fb7e-b28f-49a8-ab36-37df06875aba)

## Why It's Good For The Game
### Tailor Shop
I enjoy this new layout and think it looks better.
The upstairs store room is actually used now, lots of spare clothing and
materials are up there on the shelvesnow.
The mannequins are now behind a window so that the THIEFS can no longer
steal as easily!
Using a unique tailor shop key rather than shop1 is good for mappers and
coders.
### Merchant shop
Removed most of the fancy clothing items from thier store room, having
the tailor be focused on clothing is good- as compromise I gave them
more hoods, cloaks, and a few mannequins aswell.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->