Fixymincy (#753)
## About The Pull Request

Cleaning up some food related issues. Keeping it a draft for an hour or
something to keep collecting reports

* Some premapped base item minces yeeted
* Cheesecakes already worked as they should. For cakes, filling first,
glazing second. adding raisins tho
Their desc hints what you need. If I do a follow up I might make the
cake code better and add topping and modularity but this will do for
now.

- closes https://github.com/Vanderlin-Tales-Of-Wine/Vanderlin/issues/750

- closes https://github.com/Vanderlin-Tales-Of-Wine/Vanderlin/issues/752

* Expanded the skill quality descriptors a little. Now theres a
difference between skill 4 and 5 so leveling up and master chef makes a
difference, even if slight.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->