(BALANCE) Buff Man-at-Arms 2: The Buffening (#656)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Buffs both archetypes (Unarmed, Melee for both, Athletics buffed by 1
and Satchel given to Pikeman), undoes and changes their intro text to be
more akin to their previous incarnation (Why the fuck was this changed
in the first place?). Also gives them the Dungeoneer music as a
placeholder, they were never intended to have Garrison music, even from
the start.

## Why It's Good For The Game

I think Man at Arms SUCKS!

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: tontyGH <39193182+tontyGH@users.noreply.github.com>