(FIX/BALANCE) Buffs Jailor (#649)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Makes Jailor get Expert wrestling / unarmed at minimum, 3 in Athletics,
3 Cooking, 2 Medicine. SETS STRENGTH TO 5, INSTEAD OF JUST HAVING A 1/3
CHANCE OF ROLLING 5 STRENGTH - this is to compete with Horcs and other
crap.

Fixes some grammar issues with their intro text, adds a little to it
aswell.

## Why It's Good For The Game

Needed a buff to compete, people rarely pick him and don't understand
what he's meant to be - the Captain's helper, the prisoner's warden, and
the Garrison's grandpa.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->