[BugFix] Fix legcuffs not working at all (#783)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Fixes being unable to cuff the legs of people, that simple. For some
reason it broke a while back and nobody noticed besides a few people
(such as me) so I fixed it.

You can now target a leg with a rope/chains and tie thier legs to slow
them down once again.

## Why It's Good For The Game
Fixing broken features is good because the feature is no longer broken.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->