(MIGRANT) Adds Daywalkers, Bane of Deadites (#661)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Adds Astrata's Daywalker, a combat migrant-wave role that's meant to go
around getting into scuffles with deadites, vampires, and anyone who has
red eyes. This is a reference to Blade 3: Trinity.

## Why It's Good For The Game

Might be overtuned, might not be overtuned, it's got a low weight so i
(fucking hope) that means it's rarely going to happen. One per round,
and not invincible.

![literally
blade](https://github.com/user-attachments/assets/9b3db384-2b8a-4726-9c1d-d63abfd14ea7)


## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->