Meathook butchering reward refactor (#762)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Includes legendary skill level in butchering reward calculation, cleans
up the code.
Changes the chance to get botched result from 50% at levels 0 and 1 to
50% at level 0, 25% at level 1
Changes the chance to get perfect result from 10% at level 3, 50% at
level 4, 100% at level 5, 0% at level 6 to 25% at level 3, 50% at level
4, 75% at level 5, 100% at level 6.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
Now legendary butcher gets the same result as an average. This PR fixes
it and makes the code three times smaller
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->