Storytellers: Or how the gods started meddling in our affairs (#696)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

Its more fitting then a game_mode system that only spawns 1 antag and
then calls it quits.

The god telling the story changes based on the current highest devotion
power, so the more worshippers of a god means the better chances its
telling the story.

Each god will have their own intervention events which are storyteller
specific events that only they fire so their is some boons and malus for
each god.

Villians are added up to a cap based on living town members with royalty
and guards adding extra value.

Omens were moved to the new system so that now each omen increases the
speed at which omens get called, and omens are a unique track on the
storyteller.

Orignal storyteller from [Camper12345](https://github.com/Camper12345)
and [NanakoAC](https://github.com/NanakoAC), modified storyteller
subsystem from Horizon by Azarak (azarak908 on discord) and Burger.
Credits were grabbed from the Bubber port of Storytellers as the
original repo no longer exists.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->