Adminned clients will not get the death hud, view range adjustments (#865)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
the implementation is a bit scuffed so if you have some suggestions on
how to make it better do make a code review

- more options for `change view range`
- ghost dead hud will not show for adminned people

<details>
<summary> before </summary>


![image](https://github.com/user-attachments/assets/d278055b-b7e0-4a8d-b1c4-570d8c541d18)

</details>

<details>
<summary> after </summary>


![image](https://github.com/user-attachments/assets/83b41248-49f7-425a-89d7-29da356344b4)

</details>

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
I think it makes it easier to see what you need to admin
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->