Smells and Scents (#793)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

This PR adds smelling (inspecting with nose targeted) of reagents from
open containers!
You must either hold, or be next to the container.
The smell will be the same as the taste, unless a different smell is
specified in the object's scent_description.

The apothecary has a better sense of smell, and can identify liquids
from their scents.

## Why It's Good For The Game

I think that you should be able to give your bottles a sniff before
drinking them.
_Is this wine or lifeblood?_

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->