Adds trait Forager to identify berries (#730)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
This PR adds a new trait called Forager, it adds the ability to see a
small hint message when examining berries to some of the jobs that would
make sense to have that knowledge, jobs that don't have the trait
experience no change and no message. I added it to the roles that made
sense to me, might have missed a few or might have a few that shouldn't
have it. This trait can be expanded to cover more knowledge and perks
about the local flora. This change should not conflict with farming
level 3 giving the same perk.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
Makes it a bit friendlier to people that have colorblindness that are
playing roles that would make sense to have the knowledge of local
berries while also negating the RNG side of berry tasting. This might
even lead to additional interactions of asking someone who has the trait
if something is safe to eat.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->