Adds dungeon entrance to Malums Anvil (#869)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

* Adds small gate to the Keeps dungeon
* Moves the dungeon entrance to malums anvil
* Places the camp "Malums Kindel" next to the dungeon entrance, a small
camp for adventurers to rest in.

## Why It's Good For The Game

AHHHH!! I dont even know anymore man.

Was supposed to move smth, but now its just a new dungeon gate on malums
anvil because it was aperently missing one.
Having one there is nice. Less reason to go into the keeps dungeon.

Adds a small camp infront of the dungeons entrance, Malums kindel, (name
aint written anywhere, thats my headcanon name for it. So dont be
upsetty if you dont like it)
A place for adventurers to socialize and rest before climbing down into
the dungeon.

Lots of fantasy storys have the idea of dungeons being a puplic thing
that atracts tourists and adventurers from avar.
To the extend that people sometimes build camps at the dungeons entrance
or even in the upper lv inside the dungeon.
Dungeon meshi would be an example wich has traders even in the camp
thats in the dungeons first floor.

I find its a nice thing. Could even be a nice place for crafters to
chill and make small shops at.

Also, a corncern was raised about the keeps dungeon allowing liches to
flood the keep easily.
So the little groove of the dungeon entrance in the keep, has now a
pretty basic gate with manor lock. Just to somewhat more easily controll
who can go in and out of the dungeon.

Discord: NHC

# Testing Photographs and Procedure
<details>
<summary>Screenshots & Videos</summary>

Malums kindel and dungeon entrance

![grafik](https://github.com/user-attachments/assets/b714ec53-8c17-4858-8303-212e66ed7f07)

Itty bitty gate

![grafik](https://github.com/user-attachments/assets/b31be988-3f49-43a7-a4a7-67773ab7427d)


</details>

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->