Adds forgotten improvised surgery tool defines (#962)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
As the title says, embedding behavior was missing defines for both
retractor and hemostat for their improvised counterpart, while sawing on
the other hand was missing a define for improvised saw.

(I just realized my commit history is a mess)
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
Bugfixes are good
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

Before Changes:

https://github.com/user-attachments/assets/24221ba1-5bea-40f6-849b-8820afc2d0a0




After changes:

https://github.com/user-attachments/assets/220d3c7e-70b3-4fae-9d37-8d19b336d130

---------

Co-authored-by: tonty <39193182+tontyGH@users.noreply.github.com>