MatthiosMapFixes (#937)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## This update fixes MAJOR issues with the dungeons. 

To cover what I did

-Dungeons were removed of all simpleai apart from archers
-All mammon loot was removed apart from randomized loot pools which
still can have them spawn in.
-Dungeon floors and stairs were checked for errors and were fixed when
found
-Dungeon errors on NPC/no_equipment skeletons not equiping items were
fixed
-Dungeons were made harder to match with loot reward
-Var edits to VERY specific NPC/no_equipment skeletons to assist with
difficulty required for loot
-Addition of more NPC/no_equipment with unique loot from rooms previous
magic items.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
This PR will fix many of the issues currently plaguing the dungeon and
force adventurers to sell loot rather than take the mammons in easy to
deal with rooms. It also adjusts difficulty for specific dungeon tiles
that reward riddle of steel.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [X] You tested this on a local server.
- [X] This code did not runtime during testing.
- [X] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->