Haubergeon and iron maille color fix (#934)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

* Renames the Haubergeon, back into the haubergeon.
* Changes the Haubergeon to use the steel color pallete, instead of the
iron color pallet for the item sprite
* Changes the Iron chainmaille, to use the iron color pallet, for the
item and onmob sprite
* Gives the Iron maille sleeves
* Changes a random singel iron chain glove, dir sleve, to use the iron
color pallet instead of steel

## Why It's Good For The Game

Visual consitency is good.
And we got consitent color pallets for a reason.
Best to make sure, we dont get too sloppy with em early on.

Some uppsy seeme to have happend during porting wich cause the Iron
chainmaille, and haubergeon sprites to be swapped.
And in turn the iron onmobs seem to have been changed in respons.
Stuff that happens in a codebase after stuff travels between 4 repos, in
about a year.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->