Updates the Vampire Lord Mansion (#892)
## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

On the tin, makes something sub par better


![manor](https://github.com/user-attachments/assets/b941ebee-4f09-4752-a342-11d07dd5e97a)


## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

We are looking into adding thralls as a support class for the VLord,
this adds in some more stuff to do in the mansion proper and gives
vampires stuff they probably should have on hand. Like needles to keep
their blood bags alive!

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->