Dipping knives into poison now takes 10 units but only injects 0.5u per slash (Also lets profane dagger dip in non bottle containers) (#943)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Currently other poisons like stamina and energy poisons are rarely used
due to needing to be ingested to take effect, and as much as I want this
to be strong. The current values for injecting lethal poison (2 for
bolts and arrows, 3 for darts) take account to the slow effectiveness of
the poison. and it's strength.

With these changes the time you inject all 10 units you've already
stabbed your target 20 times, which to unarmored makes it so you already
did a crit, or are now being bashed to death by the heavily armored
person you engaged like this.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
For the methodical killers that aren't gonna slap poison straight into
the knife, having a stamina disabler allows you to nonlethally take down
someone without A. bashing their head in B. relying on wrestling vs the
current wacked out wrestling spread on jobs like most if not all serfs
(seriously why is it so random rn?)

As for the profane dagger fix, the pr is related so I decided to (try)
to fix it, though its still broken with bottles due to bottles having an
overidden object attack, and profane dagger has a overidden pre_attack
and I'm not savy enough to think of a proper flag or a sane workaround.
Sorry!
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [X] You tested this on a local server.
- [X] This code did not runtime during testing.
- [X] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->



Code in action

https://github.com/user-attachments/assets/5f68fabe-12c8-4b24-8f03-e52f1b3dc2e2

Profane bug

https://github.com/user-attachments/assets/9370932b-4538-45ef-9911-b6394e07cd74