Beauty update (#933)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

inspired by https://github.com/Rotwood-Vale/Ratwood-Keep/pull/712 , but
all code is original

- Sheltered aristocrat heir is now beautiful

![image](https://github.com/user-attachments/assets/c6018b72-f157-4f2f-b4ac-073524b9af4b)


![image](https://github.com/user-attachments/assets/6d264e32-e9de-49c3-bf20-2c6f6dceb1c6)

- inbred heir is now UGLY 


![image](https://github.com/user-attachments/assets/15401967-0cc7-46b2-bea6-ae6d0c3edc0a)


- Eoran acolytes and templars are also beautiful

- also while I was there, refactored pronoun displays and
brother/sister, son/daughter stuff instead of progeny and sibling
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
peak storytelling and rp potential
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->