MatthiosAddition (#944)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

![image](https://github.com/user-attachments/assets/7b3d2a29-c1db-45c7-982e-e6a9ad98bf59)

![image](https://github.com/user-attachments/assets/2eab0e99-7263-4047-9036-2681cf9522c2)

New addition to the dungeon, my last large tile for a while. This
dungeon will be a dedication to the adventurers when it rarely spawns in
due to the size of the dungeon. The tile hosts a large amount of loot
for a great amount of danger. At the very end there is a puzzle that if
solved while scouring the dungeon will reward a riddle of steel. This
dungeon is meant to be a war zone hence why skeletons and orcs fight.
Players are supposed to use this to their advantage.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Adds more diversity to the dungeon and a reason to explore every inch
of the matthios dungeon in order to find this large tile that may spawn
in or not.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [X] You tested this on a local server.
- [X] This code did not runtime during testing.
- [X] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->