(BALANCE) Estoc Rebalancing (#981)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Estoc's thrust now only has 30AP and gives the 50AP attack to Lunge.
Removes the chop intent as it is meant to be an edgeless blade and adds
a normal sword thrust to its one-handed intents. It is now easier to
dodge as it is already a good weapon against armor users and it is now
just as good as other swords at parrying. (why was it ultimate parry in
the first place?)

## Why It's Good For The Game

Estoc is a sword that has all of these

    Highest AP in swords, only bested by flails and impale warhammers
    Ultimate Parry
    Very Hard to Dodge
    Only one with a spammeable high AP attack

It wasn't balanced at all and it showed. Now the estoc is still the best
AP sword, but it is no longer the best sword or best weapon in the game,
it serves its purpose as an armor piercing weapon without being
overwhelming.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->