Flavor text and headshots (#963)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

This is a port from Azure Peak

gifs allowed!


https://github.com/user-attachments/assets/d9501ade-71d5-4557-88a5-45288cbd13d0


![image](https://github.com/user-attachments/assets/8e2e59b8-6613-487f-ae63-a24400142f89)


![image](https://github.com/user-attachments/assets/9d9d74a6-4b4c-4172-83d1-8a4edf24b86c)

only patreon members can add a headshot

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
FLAVOR TEXT
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->