Noose & Gallows Update (Fixed this time) (#1003)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

- Added a slapcrafting recipe for lantern posts, which can have lanterns
added and removed from them, and can have ropes tied onto them to create
gallows. This recipe cannot be built on other street lamps or lantern
posts.
- Added a mapload argument to lantern post initialization, so that
crafted lantern posts didn't spawn with a lantern attached.
- Ensured lantern posts had attached lanterns before attempting to have
them burn out, to prevent null references.
- Fixed the offset for being hung via gallows, so your character now
actually has their neck in the rope section of the sprite.
- Fixed slapcrafting rope steps also accepting chains, so objects like
nooses no longer accept chains in their recipes (fixed recipes to match
sprite).
- Adjusted difficulty of noose craft to reflect reality.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
- Craftable lantern posts and gallows provide more variety while not
majorly affecting game balance.
- Simple offset/recipe fixes make the game look better.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->