Add access and lock mapping helpers (#1009)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Basically port a modifed version of airlock helpers with only locking
and access functionality. Add defines for the access helpers to use.

Helpers are split into Town / Garrison / Church / Manor with all
required subtypes added I believe. Currently the shop / house /
apartment / waterfront keys helpers are not implemented.

IMPORTANT: These are not ready for use as there is cleanup to do with
lockids on the vanderlin map due to unused keys and unused lockids,
which I will PR after this.

I tested this with lockpicking and it seems that the init for the lock
needs to be in lateinit to make sure its after access is applied by
these.

Its possible these would have a small impact on init time.

Left to right: Town, Church, Garrison and Manor.
Below various examples.

![image](https://github.com/user-attachments/assets/b876bf9d-651c-4ab0-9267-c904fcb21abf)


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
Convience for mappers.

Varediting lockid and locked is bad for clarity and future changes.

Extendable to allow for open viewports, broken / damaged doors, open
doors, etc. For more mapping freedom.

Keys and access helpers can use the same access defines to avoid needing
to shadow changes.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->