Fixes sleeping with a good mood, letting you be inspired by two things in dreams (and adds a singular dream point) (#949)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

As the title says this is a bugfix, currently stress hard limits itself
to 0, i set the limit to -30 so the stress median code can function for
sleeping.


![image](https://github.com/user-attachments/assets/0cdb4fcc-860c-400e-84dc-e1ef566d3113)


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
Bugfixes are good
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [X] You tested this on a local server.
- [X] This code did not runtime during testing.
- [X] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: tontyGH <39193182+tontyGH@users.noreply.github.com>