Snowy additions and mapping QOL (#1048)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
- Adds missing rotations for a table type
- church/line's layer changed so items do not appear under it
- Wood floors allowed to smooth with turned floors. Also changes
icon_state for editor-ingame parity
- Fixed some grass smoothing issues
- Skeleton mobs show up as skeletons in editor
- Wall window turfs show up in editor properly

- Adds snowy turfs. sprites by what/ sgtmanguy on discord
- Adds snow tundra bushes
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
Snowy turfs and plants for Rosewood. Making the editor show as close as
possible to what will appear in ingame saves frustration and leads to
more informed design decisions. Also, placing a bunch of walls before
realizing they're windows is annoying AF.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->