UI improvements (#1049)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

This PR does 3 things improving overall UI and UX:

1. Increased the text size when examining SCOM laws, so you don't have
to squint to read them.


![SCOM](https://github.com/user-attachments/assets/56d940f0-a229-42c6-b9bc-3bf377b8829d)

2. Increased height on the following UIs: Goldmask, Purity, Peddler,
Hole of Submission, SCOM decrees and Master of Nerves, as they were all
comically low, so you had to either always resize them, or constantly
scroll through them.


![GOLDMASK](https://github.com/user-attachments/assets/9d4f8eb3-2a42-4e31-b208-7407d5e047aa)

3. You can now actually close the Migrant UI by clicking X on the
browser, and it will stay down and not constantly pop back up. If you
were part of the migrant queue, it will remove you from it with a
message, so nobody gets drafted if they would accidentally forget to
manually leave the queue before closing the menu.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game

Better UI and UX.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->