No more lockhash and custom key update (#1055)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Remove lockhashing completely, only use lockid.

Kill the custom_key_blank item at tonty's request functionality moved to
key/custom.
Some notes on this; 
keys don't have lockids until they are finished and can't be modified
after that outside of their name, it should be impossible to have a key
with a null lockid that is usable.
Custom key lockids are strings from number input, therefore no doors or
keys should just be using stringed numbers unless explicitly intended to
be opened via making a custom key as part of a puzzle.

Add feedback for custom keys and locks (lock examine code is bad...)

Remove closet emag act because what the hell is an emag.

Remove door trypicklock because its replaced with the skyrimed
lockpicking.

Closets can't be locked if open like doors.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
Unneeded system that can't handle updating locks midround AND made init
longer by a not insignificant amount.

custom_key_blank was bad.

Feedback for making custom keys and locks good.

unused code bad.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->