Meatpie smite (#902)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Adds itembound to component, allowing players to be trapped inside items
(used for the smite, ported from monkestation).
Adds transformation_animation, allowing for animations to be played
while people are transformed(used for the smite, ported from
monkestation)
Adds an adminsmite to turn sinners into a meat pie.
## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->
Adds an admin smite to turn someone into a meatpie. Drag asked for this,
don't ask me.
It's probably fixed this time.
## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: tonty <39193182+tontyGH@users.noreply.github.com>