Gives the Apothecary Master Alchemy, removes lockpicking and pickpocketing from their skills. (#1182)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Gives the Apothecary Legendary Alchemy
Remove Lockpicking and Pickpocketing skills from the Apothecary

## Why It's Good For The Game

The Apothecary that has the main job of making potions has their highest
skill as pickpocket(Master), meanwhile the Court Mage has Master Alchemy
, just seem to me that someone that spent most of their life working as
a potion maker is also VERY good at it.
But they also got pickpocket and lockpicking skills that i can't see any
relations to the job or these skills being actually useful, the
apothecary is very rich, they don't need to break into places or
picpocket people in the streets nor are the jobs of an Apothecary, also
the description don't gives any clues why would they know that.
In Short, removes two useless skills of the Apothecary and increase
their main skill that it is used to do their job.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->