Gives leylines absorb to archivist and magic book to the magicians that were missed. (#1186)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

closes: https://github.com/Vanderlin-Tales-Of-Wine/Vanderlin/issues/1073

Gives leylines absorb to those magicians that were missed such as,
Archivist, Heartfell Mage and Dredger Mage roll, also give them their
respective books.

## Why It's Good For The Game

Much like any adventurer these mages have acess to the bare minimum
besides that spawning as a magician without the basic items generate a
very boring gameplay to catch up to everyone, chalk, plants stuff like
that is easy to get and fast, but things as baptism and making a book
can take a long time.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: grungussuss <96586172+Sadboysuss@users.noreply.github.com>