Updates Several job tutorial tooltips, removes merchants nobility, makes the sheltered heir and inbred heir virgins (#1221)
## About The Pull Request

This PR does a few things:

Least popularly, it removes the merchants nobility. This has been
discussed in depth with project leads and other roles are on the
chopping block when other content is ready. Merchant does not need to
wait.

This PR makes the Sheltered Aristocrat and Inbred Wastrel virgins for
Cult sacrifices. Daring twit has not been touched.

Several jobs had their tooltips updated, including:
<details>
<summary>Monarch</summary>
<br>
"Elevated to your throne through a web of intrigue, political
maneuvering, and divine sanction, you are the unquestioned authority of
these lands. The Church has bestowed upon you the legitimacy of the gods
themselves, and now you sit at the center of every plot, and every
whisper of ambition. Every man, woman, and child may envy your power and
would replace you in the blink of an eye. But remember, its not envy
that keeps you in place, it is your will. Show them the error of their
ways."
</details>
<details>
<summary>Royal Knight</summary>
<br>
"You are a Knight of the royal garrison, elevated by your skill and
unyielding loyalty. Sworn to protect the royal family, you stand as
their shield, but your loyalty is not blind. You are bound to serve
them, but should the monarch forget their place with the gods and the
land, you are not above breaking your oath to restore the balance. Or
are you?"
</details>
<details>
<summary>Town Elder</summary>
<br>
"You were once a wanderer, an unremarkable soul who alongside the
Veteran, Matron, and Court Magician, carved your name into history. Now,
the days of adventure are long past. You sit as the Towns beloved Elder,
while the crown may rule from afar, the people look to you to settle
disputes, mend rifts, and keep the true peace in town. Not every
conflict must end in bloodshed, but when it must you will do what is
necessary, as you always have."
</details>
<details>
<summary>Court Magician</summary>
<br>
"A seer of dreams, a reader of stars, and a master of the arcane. Along
a band of unlikely hero's, you shaped the fate of these lands. Now the
days of adventure are gone, replaced by dusty tomes and whispered
prophecies. The ruler's coin funds your studies, but debt's both magical
and mortal are never so easily repaid. With age comes wisdom, but also
the creeping dread that your greatest spell work may already be behind
you."
</details>
<details>
<summary>Matron</summary>
<br>
"You are the Matron of the orphanage, once a cunning rogue who walked
the shadows alongside legends. Time has softened your step but not your
wit, thanks to your kinship with the Veteran, Court Mage, and Town
Elder. Now, you rule with a firm, or gentle, hand. Ensuring the orphans
under your care grow sharp, swift and self-sufficient. One dae, they
might just be as skilled as you. Maybe a thieves guild legend you once
were."
</details>
<details>
<summary>Veteran</summary>
<br>
"You are a renowned adventurer, once hailed as a savior of these lands,
at least while everyone still remembered. Alongside the Matron, Court
Magician, and beloved Town Elder you rose from nothing to everything in
the blink of an eye. Nowadays however, your bones ache, you've retired
your sword, and the glory has faded. You wonder what your old friends
are doing."
</details>
<details>
<summary>Merchant</summary>
<br>
"Born a wastrel in the dirt, you clawed your way up. Either by luck or,
gods forbid, effort to earn a place in the Merchant's Guild. Now, you
are either a ruthless economist or a disgraced steward from distant
lands. Where you came from no longer matters. What matters now is you
make sure the fools around you keep buying what you sell. Everything has
a price."
</details>
<details>
<summary>Archivist</summary>
<br>
"A well-traveled and well-learned seeker of wisdom, the Archivist bears
the mark of Noc's influence. Tasked with recording the court's events
and educating the ungrateful whelps the monarch calls heirs. Your work
may go unappreciated now, but one dae historians will sing of your
dedication and insight."
</details>
<details>
<summary>Butler/Maid</summary>
<br>
"You are the master of the household staff, ensuring that meals are
served, chambers are kept, and the court is kept clean. Though you wear
the royal colors, you hold no true authority. A servant among servants,
yet without your guidance chaos would reign in the kitchens and halls."
</details>

 

## Why It's Good For The Game

These are mostly flavor and tone shifts as discussed with the rest of
the lore project leads. We want the preface that the Matron, Veteran,
Court Magician, and Town elder have some vague history that could still
be ignored for player autonomy. The merchant's nobility is superficial
and unneeded along with a couple other roles that will be adjusted in
the future. Sorry besties you need to pay taxes now.

The two heir subtypes were given virginity to encourage cultists to
possibly target the royal family for their rituals. Powerful but risky
additions.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: grungussuss <96586172+Sadboysuss@users.noreply.github.com>