Fixes steamknight boiler overlay. (#1354)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
As it says in the title. 


![image](https://github.com/user-attachments/assets/d91f8f7b-1ab0-4ab8-939e-3aedf4b87282)

![image](https://github.com/user-attachments/assets/b0df3462-422b-4382-b192-6192bc8a1e68)


<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
When creating the sprites, I had expected an implementation where they'd
layer behind the armour when facing South. I was incorrect, this makes
them look as they should.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

Co-authored-by: AberrantQuerist <crystalseeker04@gmail.com>