Gaffer related changes (#1319)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

-removes MGE line from mercenary medal desc since MGE doesnt sign the
tokens anymore

-adds a merc guild ad to the adventurer post_equip since Gaffer is also
related to adventurer work

:"If I wanted to make mammons by selling my services, or completing
quests, the Mercenary guild would be a good place to start."

-gives gaffer dodge-expert trait, changes 2 int buff to speed 2, buffs
bows, crossbows and swimming by 1 point, with more buffs to stats if the
gaffer is old (crossbow, bow, unarmed, knives, swimming climbing by 1,
stealth and athletics by 2 and gives another point to perception) since
they are supposed to be a ranger in lore

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

its only natural

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->