Mannequin Update For Wrists, Neck, Ring Display (#1363)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Updates mannequin code to allow for rings, wrists, and proper neck slots
to be displayed. The neck intent now correctly equips the neck instead
of the cloak. Also did some small refactoring of the menu UI (the close
button should work now) and added examine punctuation.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game

It really, really bothered me that I could not equip the bevor or
vambraces on mannequins. This PR corrects that, allowing for wrists,
neck, and even a ring slot. Rings have next to no on_mob sprites though,
so they are more of a future proof thing. As a consequence, rings and
cloaks have to be equipped via the UI menu because there is not a
corresponding intent for them. More slots can be added, maybe for
backpacks specifically, but I felt that this was enough to properly
display armor/clothing/jewelry.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->