(BALANCE/FIX/ADDITION) Dredge - the unfuckening (#1288)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

- (Hopefully) Gives the Dredge the CORRECT bullet pouch.

- Changes armor roll stats around for a (slightly) better statspread
chance.
(Black Knight statspread notably buffed)

- Adds new 'tarot' roll, with a random meaning and (generally) slightly
better stats.
(This uses pickweights incase they need to be balanced further. Might
not, but an added dial helps.)

- Lowers (with the exception of Kobolds, raising some stats by 1, and
with the exception of half-elf and drow being lumped in with other races
apparently) all races major stats by 1. This is to make room for the RNG
rolls.

This might be overtuned now but if it is some negative stats can just be
added onto the new 'Tarot' roll.

Dredge now isn't as shit to roll. Net benefit for the people who
actually enjoy rolling the virtual slot machine.

<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->