Fix char appearance randomization (#1160) (#1255)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

This PR is an attempt at fixing #1160 , which has terribly vexed me for
quite a while. To that end, it makes adjustments to the appearance
randomization code while trying to respect antag randomization.

Some players didn't experience #1160 because, presumably, their
savefiles were written before appearance randomization was taken out.
For players that do experience this issue, saving and reloading your
character doesn't work due to the game loading the `randomise` list,
which is initialized to randomize everything.

Because there's no user-facing options to adjust those randomization
parameters, I thought it might be a good idea to turn off most of it for
now -- the original code is either still present as a comment or trivial
to revert. I tried to respect the antag randomization (though the code
doesn't do much in that regard, from what I can tell).

## Why It's Good For The Game
I doubt that forcing certain unlucky players to have a random appearance
was intentional. Bug fixes are always good!

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: grungussuss <96586172+Sadboysuss@users.noreply.github.com>