Templar/Paladin helmet and misc. changes. (#1215)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
This PR changes:
-Introduces a new "generic" helmet that some of the patrons with no
unique sprites have called "Bastion Helmet" to make it easy to identify
as it's a unique helmet with no crafting recipe.
-The helmet Templars and Paladins get by default, moving the stats from
the subtype that the ones without a unique sprite had before and gives
it to the 'necked' version that most use as a subtype.
-Puts the Ravox version of the helmet on Ravox Paladins which they
didn't have before.
-Deletes some of the redundant flags the helmets had before.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
Fighting as a Templar or anyone using that helmet sucks, honestly it
just sucks, it feels like you are playing with tank controls as you
can't even hit someone directly diagonal of you most of the time and
your awareness in combat is basically 0, most of the time it's just a
better idea to take the helmet off and risk the head crits than used
that cursed tin can that you get.

![dreamseeker_j9nTZrl5TQ](https://github.com/user-attachments/assets/3f4f1411-a439-4570-b351-7272733c7a3e)
The code for the helmets is in all honestly a bit all over the place,
this brings it a tiny bit closer to being more consistent while also
helping with Templars not rushing the smith to get a different -and
actually useable- helmet.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: tonty <39193182+tontyGH@users.noreply.github.com>