Steward stat changes (#1457)
Faster but dumber

## About The Pull Request

Steward gets the -2 SPD debuff removed, but there +8 INT goes down to +5
INT.

## Why It's Good For The Game

The stewards entire job is running around and managing the economy,
which is AGONY when you have 8 SPD at your best, 7 when you're drunk or
hungry or thirsty, 6 when you're tired, etc. To counter this, I lowered
the INT by 3, the steward still gets great INT, but not THAT great. It
doesn't make sense that the coin counting guy is smarter than the
archivist, who knows every language and various spells.

## Pre-Merge Checklist

- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.