Post-S&S Rosewood (#1167)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

This adds Rosewood and also makes some of the use of the word
"Vanderlin" in reference to the town ambiguous, using the relevant town
instead. Map done by TheDragMeme, Halfcool, and Chair Raptor!
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [ ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: HALFKOOL <Austinle@mail.com>
Co-authored-by: tontyGH <39193182+tontyGH@users.noreply.github.com>
Co-authored-by: Thedragmeme <thedragmeme@gmail.com>
Co-authored-by: Raptorian1 <maxheydorn03@gmail.com>
Co-authored-by: grungussuss <96586172+Sadboysuss@users.noreply.github.com>
Co-authored-by: CheffieGithub <113442598+CheffieGithub@users.noreply.github.com>