Deity Influence System (#1431)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

This PR, in short, does 3 interconnected things:

- Replaces old statistic names with defines, to prevent mistakes and
improve readability
- Adds 42 new statistics
- Adds Deity Influence system

Deity Influence System is quite straightforward - Each God has a set of
influences (ala statistics), which may be positive or negative for the
given God. Sum of these influences is their total influence, which could
be used in the future used to pick storytellers, once the system is
properly tuned. For example, Necra has "Graves consecrated" influence,
which means that with each grave consecrated, her influence grows by the
weight given, up to the limit for that given influence. On the other
hand, she also has "Graves robbed" influence, which does the opposite -
her influence lowers with each grave robbed.

Example pictures:


![InfluenceP1Final](https://github.com/user-attachments/assets/5071fdcb-6141-4041-8288-137e614f15e3)


![InfluenceP2Final](https://github.com/user-attachments/assets/a51d693c-97f4-43cf-a27b-8c5461b21c1e)

Influences are visible to everyone at the round end, and I have also
added "View Influences" button to the Gamemode panel, accessible only to
admins during the round. Definitely handy for event makers, as they can
for example see, that briars made many sacrifices to Dendor, and that
they could make some interesting RP with them regarding it.

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

New statistics are certainly useful and interesting for players and
admins alike. Influence system can be further used with storytellers to
make certain events more likely, or to make more dynamic storyteller
picking.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->