Royal Knight armor upgrade. (#1467)
Changes probability of getting a knight helmet to 100, and gives the
royal knight chain gauntlets on round start.



## About The Pull Request

Gives the royal KNIGHT the KNIGHT helmet, and chain gauntlets.

## Why It's Good For The Game

This is good because currently, the royal KNIGHT, who is a NOBLE, and
PERSONALLY RESPONSIBLE FOR GUARDING THE ROYAL FAMILY, has completely
exposed hands, and a 70% chance of having completely exposed face. This
PR rectifies this by giving them gauntlets and guaranteeing the knight
helmet. It doesn't make sense for a knight to not start with gauntlets
but a MAA does, and the knight helmet is CALLED the knight helmet, the
knights should have it no matter what.

## Pre-Merge Checklist

- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.