Gives pilgrim physicker the physicker combat music (#1494)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

I copied the line of code from the physicker that gives them their music
and added it to the pilgrim physicker.

## Why It's Good For The Game

This is good for the game because the music is cool and more people
should be hearing it. Also if you think about it the pilgrim physicker
plays more like Pathologic than the townie one. Also, and more
importantly, it is good for the game because I personally say so.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->