[MDB IGNORE] Implements a stat modifier system (#1388)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Adds the ability to apply sourced stat modifiers. 

## Why It's Good For The Game

Systems that want to apply stat modifiers have to do so with a single
shared variable, which causes large amounts of conflicts and
headache-inducing loss of state. This method will make working with
these systems much easier.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->