High Collier | Port from Azur (#1412)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Sprite port from:
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/1873

* Adds iron High Collier
* Adds steel High Collier

## Why It's Good For The Game

Sprite port.

Gives a sidegrade to the chaincoife. A High Collier, either from iron or
steel.
It is sligthly stronger then a coife, but in exchange does not cover the
head.
You coul also call it the ligth version of a bevor.

Its a low impact side grade, looks cool, and gives some varity in armor.
Nice thing.

Sprites by Cre, commissioned by Witty.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: grungussuss <96586172+Sadboysuss@users.noreply.github.com>