Adds a new Troll variant: Cave Trolls! (and also cleans up troll code, oh and also dendor taming chance) (#1349)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Adds a new cave troll variant: The cave troll. this variant comes with
normal variant and "hidey" variant. similar to that of regular trolls.
the major differences between these cave trolls and normal trolls is
it's unique stone throwing ability.

These trolls spawns in the cave areas of Malum's Anvil. replacing the
orc spawn there. as it would be more thematically fitting.
(also increased the spawn rate a bit, for fun :) ) and the ambush trolls
spawns as did bog trolls. just now on caves.

Oh and they drop mana crystals too, as compensation for the harder
difficulty

Here's it in action:


https://github.com/user-attachments/assets/cd0315da-3709-497e-a663-48b15a63b5cb

Also cleans up troll code a little (WHY IN THE EVERLOVING FUCK WAS TROLL
AND BOG TROLL TWO DIFFERENT TYPES INSTEAD OF ONE BEING A CHILD OF THE
OTHER)

So that dendors don't get to play with free cave troll arty. I reworked
dendors taming to add the possibility of probability instead of it
beaing a 100% chance to tame. and made it so that trolls only have a
good chance of being tamed and cave troll a low chance
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
More enemy variations is fun! and it's not just a reskin of normal
trolls. it has something that differentiate it with normal trolls. Makes
it more fun to fight with instead of just keeping your distance and
bleed it to death
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->