Buff and clean up transfix (#1536)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Vampire lord starts with SKILLED (3) blood magic. Still goes to master
on level 2. This change is because vampirelord struggles to transfix
people since now after their blood magic went from 4 to 2 when vampire
was fixed.

`transfix/master` combined into base. Cost reduced to 150 (50% increase
instead of 1000% jesus christ)
Base blood magic die changed to 9. Willpower divisor reduced to 3.3,
this is to create more fair rolls slightly in favour of vampires in the
majority of cases.

Adjust drowsiness boundaries to allow for typing time.

Combat mode bonus to willpower changed to an extra die roll.

Arcane/Holy now does a roll to see if they expose the vampire instead of
just checking if either skill is 1 or above.

Fix transfix not returning TRUE when successful.

Fix runtimes in case of target being an NPC.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->