Deity Influence V2: Influence now picks the storyteller, adds 6th influence to each deity (#1476)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

This PR does 3 things regarding god's influences:

- Storyteller is now picked by the deity with the highest influence,
instead of just the one with a highest number of followers. Followers
still serve as an important source of influence with a high point gain,
which can be even customized for each deity. Additionally, each time
storyteller is picked by the influence, the highest and lowest
influential deities receive minus and bonus influence point,
respectively.
- Adds 6th influence to each deity, to help balance influences, list
below.
- Minor fixes and balancing for existing influences based on the rounds
played. Some changes to the stats themselves, like making laws made also
count decrees, and exchanging Zizo's number of active deadites with
number of deadites woken up. Matthios also got "dodges made" stats
instead of number of kleptomaniacs, to help him out influence-wise.

It also adds the most influential deities to the statistics panel, and
modifies it a little to accomodate the changes. In that panel there
should be preferably mostly general stats, but we do not have that many
yet.

**New and changed influences:**

1. Astrata - Taxes collected, Laws made changed into Laws and decrees
made
2. Noc - Mana spent
3. Necra - Total deaths, moved from Graggar
4. Pestra - Food rotted
5. Dendor - Deaths in the forest
6. Ravox - Number of underworld duels
7. Xylix - Songs played
8. Malum - Blacksteel smelted
9. Abyssor - Water consumed, People drowned
10. Eora - Number of pacifists
11. Zizo - Tortures performed, Number of deadites changed to Deadites
woken up
12. Graggar - Blood spilt, moved from Abyssor, Limbs bitten
13. Baotha - Luxurious food eaten
14. Matthios - Locks picked, Number of kleptomaniacs changed to Dodges
made

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

Much more dynamic and interesting storyteller picking than just simple
number of followers. New stats help to balance influences and provide
insight into the round for both admins and players alike.

<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->