Chain kilts | Port from Azure (#1407)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Pr this is ported from:
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/1860

* Adds Steel and Iron kilts, a resprite of normal chain chauses
* Adds crafting recipy for Chain kilts
* Adds chainkilts to supply packs
* Adds chainkilts to loot spawners

## Why It's Good For The Game

Porting cool sprites is cool

Sprites by Dakken and Cre
Edits for on_mobs by me

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->