Town Elder role Revamp (#1205)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

This PR aimed to revamp the Town Elder role, similar to how Veteran
works with multiple classes, town elder was given a similar look but it
is more broad in terms of what the Town Elder could be, it was added in
total 5 classes.
The mayor role was kept but it was turned into a Class.

Town Elder PQ req was also increased to 10, since it requires atleast a
decent RP and knowledge for the classes.

Town Elder can now make Announcements in the Inn as requested, it now
list their name and the Town Elder title to not confuse with the
different subclasses.

Announcement:


![image](https://github.com/user-attachments/assets/a836d516-da21-44bd-9518-f3d3efc48345)

Mayor - Much like the previous role , but this time it was given
Nobility and Merchant eyes too, but it is a much weaker combat role and
more diplomatic than ever, aims to be the mediator between the crown and
the people.


![mayor](https://github.com/user-attachments/assets/4d2cd07c-dc8b-4cb0-89ae-b4ef872996d7)

Master of Crafts and Labor - A production class, Heavy RNG that can get
skills ranging from Average - Master in multiple crafting skills , not
meant to combat but to both help New players in their respective roles
and fill some holes in the lack of roles in the city, basically a
handyman.


![mastercrafts](https://github.com/user-attachments/assets/9246d066-53f8-4995-853f-cafe5da0f58f)

Hearth Acolyte - An Acolyte that left the church after saving a noble,
turning a new leaf on their life, rather than serving the church they
now walk the path of helping the people. One of the more vague classes,
more up to what someone personally wants to RP, skillset is very similar
to an Acolyte besides being more fist focused, meant to help the town
and church faction situations.


![acolyte](https://github.com/user-attachments/assets/2715854e-02a9-4698-9864-fb443f504b48)

Dreamwatcher: - A person blessed by Noc, not as a magician but as a
dreamer, no combat skills and few basic skills such as crafting, hight
INT but poor and the weakest class out of all since it has no combat
skills, with the only upside being their special skill to dream about
two things everynight, the god with the highest influence and about a
random antag that is active and alive, ofcourse none of those are given
granted and are cryptical and vague, you are also able to use it when
you are tired so it is a limited use thing.


![dream_watcher](https://github.com/user-attachments/assets/63c0eb24-9734-4e29-8ec2-dc54943932cc)

Example of a dreamer, Matthios was the one with the highest influencer
and the active antag was a bandit (some messages are more vague, others
are easier to understand, up to RNG) :


![image](https://github.com/user-attachments/assets/78081bec-8335-4ba6-bbcf-f48c27a26674)


Lorekeeper: A famous bard that took the path of recording the story of
the town, from legends to facts, worse music skills than the mayor but
more broad skills, less money, meant to be a teller more than a bard ,
helping the town through their wisdom collected through times.


![image](https://github.com/user-attachments/assets/25c2a46e-8860-4afb-919f-f89abd4822e8)

Misc : Added a keyring for 1 class and reduced the bank amount to 50
since the original mayor now get more money in their pouch and the other
classes are not that rich.
## Why It's Good For The Game

I personally found this role one of the most unused ones, most rounds
nobody picks this role unlike the Veteran and i believe that is has
great potential, the classes are weaker than the veteran so it is not
meant to overshadow them but they have their own unique touch to give
potential for great rp situations.
Hopefully this could make Town Elder a role that people actually play
and enjoy or even roll High for it, i know the classes are very broad
but i still tried to keep it as keep worker since the role is
categorized as Garrison.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: E231kei500bandai <47338680+E231kei500bandai@users.noreply.github.com>
Co-authored-by: grungussuss <96586172+Sadboysuss@users.noreply.github.com>