[PORT] Plexora optimizations (#1416)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

Ports:
- https://github.com/Monkestation/Monkestation2.0/pull/4454
- (partially) https://github.com/Monkestation/Monkestation2.0/pull/6259
- https://github.com/Monkestation/Monkestation2.0/pull/6288

Other:
- Changes the Discord invite to the preferred one so we don't have to
make a commit to change it every
time the invite changes
- Removes a few "monkestation edit" comments

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->