Decorated Halfplate | Port from Azure (#1410)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

* Adds decorated Half-Plate
* Adds decorated Half-Plate with silk Corset
* Adds smithing recipys for Decorated Halfplates

## Why It's Good For The Game

Porting nice sprites be nice.
Reskin of halfplate, so no balance problems.

Couldnt pin down sprite source.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->