Minor Fixes. (#1510)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game


Various Optimization and Upkeep ports links here:
https://github.com/tgstation/tgstation/pull/80941
https://github.com/tgstation/tgstation/pull/70966

Forgot to link this in the last pr:
https://github.com/Mojave-Sun/mojave-sun-13/pull/2512 for the framework
of bridges

https://github.com/Rotwood-Vale/Ratwood-Keep/pull/1654 < grabbed drill
sprites from here

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: tontyGH <39193182+tontyGH@users.noreply.github.com>
Co-authored-by: grungussuss <96586172+grungussuss@users.noreply.github.com>