Cursor Changes (#1605)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
I made these cursor icons after seeing our current ones in another game
(didn't write it down, don't remember but I know that it's probably not
ours to use) These new ones make the gen click charge look pretty neat I
think. There are also click drag icons aswell as hover icons for if
coders ever wish to implement that into the game. I also removed the
human_give.dmi icon, it was unused for the most part, we don't even use
give intent anymore that was gutted like 3 years ago when it was made to
give people items via right click.


![image](https://github.com/user-attachments/assets/de2b5fa3-ef13-4b84-ac1f-ce07d6575291)


## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->