New viking helmet (#1676)
This is a remake of a previous PR I made, updated for the new version of
the game.

## About The Pull Request

Adds a viking helmet, a spangenhelm, to the game. Costs 1 steel to make,
and covers the eyes, head, and nose. It's a cross between a greathelm
and an open faced helm.

## Why It's Good For The Game

Variety to our armor, too many people are walking around with hounskulls
and visored sallets. Just a splash of extra flavor. Sprites custom made
by me.

## Pre-Merge Checklist

- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.