Knockening (#1670)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Instead of immediately knocking on a locked door when it is locked and
you clicked on it, you now have to be on harm/punch intent. Otherwise it
defaults to trying the handle and the rattle animation/sounds, exactly
like walking into the door.
Furthermore I changed some wording when attacking windows and doors with
claw intent, since mobs other than deadites use it, changed some checks
for tipping over mannequins and made secret doors blend in even better
with the surrounding walls.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
It seemed weird that by default your character immediately knocks on a
locked door
it's weird that you'd get a "The deadite smashes the window/claws at the
door!" in chat if the mob happened to be a goblin, for example, or that
you'd see the assailant from the other side of a locked door.
Secret doors now have less metagaming potential, acting almost
indistinguishable from from walls at first glance. Also removed a
redundant 'the' that would be displayed twice.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->