ring servant changes and one hailer change for good luck  (#1623)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

-adds is_gaffer_assistant_job
-adds is_gaffer_assistant_job to hailer and headeater so they can be
used by the ring servant without the ring
-lets ring servant see headprice
-prevents ring servant from getting the ring via trying to pick it up
from the ground or statue
-adds ring servant's keyring type and gives it to them, removes the key
they had on their neck
-changes ring servant from -20 pq to 8
-gives ring servant math 1
-changes hailer to use .read instead of .examinate because .examinate
doesn't fucking work to read something in this codebase and its a dumb
fucking rigmarole to click the examine read instead, fuck .examinate on
paper just make it all .read

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

more people to run the counter is good, pq is needed for when ring
servant is actually the one keeping the gaffers in check.

<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: grungussuss <96586172+grungussuss@users.noreply.github.com>