Mapping stuff | Metal Torch Sconce (#1650)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

+ Adds a version of sconce that is holding metal torches

## Why It's Good For The Game

Did you know metal torches are brigther then normal ones? The radius is
1 tile bigger.
Doesnt seem like much, but remember. That measn that the circel of ligth
gets 1 tile further in each direction. Its big.

Only adds the ability for them to be placed via mapping.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->