kills the terrible dirt fix "bearerdied" var from the ring of burden with a fix that corrects the issue that it prevented (#1713)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

- Removes "bearerdied" var from the ring of burden, which was in place
to quickly fix an issue with gaffer spawning resulting in the ring of
burden destroy proc being called, sending the signal and causing a dupe
ring

- Adds visualsOnly to gaffer pre-equip so the real ring of burden isn't
used in the preview, which will be destroyed after creation and send the
signal to begin with

## Why It's Good For The Game

it kills the bearerdied var, which only creates issue. like this one
https://github.com/Vanderlin-Tales-Of-Wine/Vanderlin/issues/1333

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->