Character Flaw: Pacifist (#1617)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

This PR adds a new challenging character flaw - Pacifist!

Functionally it works the same as the usual pacifism, which you can get
from wearing eora's flower for example, except there is no way to remove
it...

The only exception is made for antagonists, as that could hinder their
role. For that reason, if a roundstart chosen antagonist has a pacifist
flaw, it instead rolls a random new one. If the antagonist is created in
the middle of the round, the pacifist flaw is supressed, and it will
re-emerge if the antagonist status is removed.

Also contains a minor character flaw refactor to faciliate the new flaw,
which could also lead to more interesting flaws in the future, or more
flaw mechanics, like re-rolling flaws in the middle of the round as
reward for something. This refactor also solves the issue where
character flaws could not access mob minds, which caused Badsight bonus
effect to not work at all and hindered development of more flaws.

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

New interesting flaw for hardcore gamers and roleplayers alike.
Character flaw refactor allows for more interesting flaws in the future.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->