Maintenance on RT related things (#1631)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Removes:
Ticker vars:
isroguefight
isrogueworld
#defines:
MATURESERVER
ROGUEWORLD
DEPLOY_TEST
FASTLOAD
Travis related defines because afaik they are not used
Files:
rt.dm
mapsrw.txt

Commented out underwear removal code
Not commented out pants pulling down code and the commented out vars for
it

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
Maintenance, we don't need these

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->