Forest Brawler becomes Forest Reaver (#1768)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->This shifts the widely controversial subclass of forest brawler to be
a more identity driven variant of forest guard. Their skills shift from
being a pugilist to a dual wielding and or two-handed axe/mace wielder
with a bit more stamina and a bit less speed than the infantry subclass,
while still having the ability to dodge as they don't focus on defense,
but rather a berserker aesthetic. Wrestling is now equal to the
infantry, at tier 3 instead of 4.
Summary
- Lowered wrestling to equal infantry
- Gave Warden actual combat axe
- Gave Reaver woodcutter axe
- Changed Brawler to Reaver, replacing Dodge Expert from Brawler to No
Pain Stun
- Swapped three strength to three endurance and made subclass malus for
speed to neutral instead, leaving it still the slowest subclass.

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->A grizzled old berserker is a better fit to the PTSD
ridden barracks of the Forest Guard than some guy who is just a monk but
better. Has more identity, and is more feasible to facing threats
without solely focusing on wrestling.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->