Icon smoothing update [IDB Ignore] (#1622)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Updates icon smoothing to a more modern standard. Most smoothing is
handled via bitmasking, edge smoothing is icky and doesn't but it's a
bit better than it was before, if a bit hacky.

The short of it is if you are a spriter you only need to fully sprite 4
(Cardinal connections) or 5 (Diagonal and Cardinal connections) states,
create a .png file of this then a .png.toml for it. Then running
`tools/icon_cutter/_run.py` will create/update DMIs based on the toml
config and png file (destructive so rename a changed file if it matters)
the icon_cutter can also be ran manually via command line.

Example input png (5th state is not required and unused in this
instance)

![stone_deco](https://github.com/user-attachments/assets/33d1203f-aa30-493e-9e5f-5f22e9e67b56)

Example input png.toml
```
output_name = "stone_deco"
template = "cardinal_32x32.toml"
```
Example output DMI

![image](https://github.com/user-attachments/assets/7a8059a5-16b7-4ea7-b401-fcf68150f7f1)

This can be done for anything that connects to something else, mainly
for Turfs and Objects like tables but could be used to make masks or
overlays as well.

Neighborlay smoothing is now handled by SMOOTH_EDGE and instead of
canSmoothWith (Or now smoothing_list) being what atoms can overlay their
neighborlay onto X its what atoms X can overlay onto, this is so the
behaviour is the same as other smoothing flags.

This is done by adding an image overlay to X and offsetting it so it
layers over the atom.

Importantly this also fixes all the incorrect smoothing flags set on
things that couldn't smooth.

Sidesteps init icon operations by using MAP_SWITCH but this has the
negative of the compile icon for walls actually being an invalid icon
state which is used to set a valid icon state by smoothing. This means
that if a map editor that is not StrongDMM is used then all walls will
be missing icons. This can be fixed by just adding an icon to the files
but it would probably get overwritten by the cutter running.


![image](https://github.com/user-attachments/assets/9dd0fc7b-1aa6-4a7f-ad92-2cc352c60a86)


![image](https://github.com/user-attachments/assets/eeac3a1c-5e7f-4084-b766-eec06dc9a0fa)

Ports
https://github.com/tgstation/tgstation/pull/52864 (And fixes)
https://github.com/tgstation/tgstation/pull/53906
https://github.com/tgstation/tgstation/pull/54129
https://github.com/tgstation/tgstation/pull/69741
https://github.com/tgstation/tgstation/pull/71989
https://github.com/tgstation/tgstation/pull/84402

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: grungussuss <96586172+grungussuss@users.noreply.github.com>