Attack animations (#1809)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
this is by large a port of
https://github.com/tgstation/tgstation/pull/88418 with modifications
from me.

<h1> <details>
<summary> PREVIEW </summary>


https://github.com/user-attachments/assets/978e7802-bf81-4321-90b2-b74a6dca2ef0


https://github.com/user-attachments/assets/d6dfe4e1-2f43-41c7-8d80-26cda119d33a

</details> </h1>
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
it looks cool
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->