Mage Attunement Changes (#1785)
## About The Pull Request
Currently Zizo Mages are unable to do most magical spells, considering
they are evil and are more likely to need offensive spells, being able
to do magic that isn't just Hydrosophy would be ideal. As most of the
schools they are proficient with by default aren't really useful at all
or are utility spells.

I've also edited Noc to be more proficient then Zizo at pyromancy.
Generally Noc Mages are still substantially better then Zizo mages, but
at least it's not a complete joke. The whole point of following inhumen
gods as a mage should be that you want power at any cost, so they
shouldn't be just bad at magic in general.

I understand if there might be a desire in future for them to focus on
different magicks, but due to the facts that a lot of their attunements
do effectively nothing to help them is an issue. They also just
generally have less attunement and are worse then Noc mages by far, even
if their attunements were actually for spells they used.

## Why It's Good For The Game
See Above.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your cxode/feature/fix
locally. -->