Mercenary balance adjustments (#1774)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
--> This brings all mercenaries and town watch into the skilled/low
expert tier for combat, allowing MAA to be the step up before knights as
far as defensive line in the town.

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. --> This prevents power creep on mercenaries end and gives
dwarf underdweller bombcrafting skill for soul purposes, removes expert
wrestling from a few mercenaries, and allows black oak elven mercenary
to actually dodge, considering it a shieldless light armor character.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->