Map-based monarch titles. (#1866)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Monarch titles are now defined in the mapping configs.
Roguetest is god/goddess.
Vanderlin is king/queen.
Rosewood is emperor/empress. (Drag's request.)
Daftmarsh is count/countess.


![VanderlinTest2](https://github.com/user-attachments/assets/52ccd44b-0991-4167-b653-406db744ce5e)

![VanderlinTest](https://github.com/user-attachments/assets/4c8230d9-001b-470f-9960-a60dcc12e1a5)

![RosewoodTest2](https://github.com/user-attachments/assets/c5fb8d23-20df-48b6-ab5a-3e96d47fecf9)

![RosewoodTest](https://github.com/user-attachments/assets/17215fe8-2488-4202-9c6a-e44656cb84fa)

![DaftmarshTest2](https://github.com/user-attachments/assets/0bb4a443-673c-4195-84f5-4a1c21ca4d4a)

![DaftmarshTest](https://github.com/user-attachments/assets/f51508d8-d6e7-40cc-a822-85c1893c0715)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
Good for flavour and lore.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: AberrantQuerist <crystalseeker04@gmail.com>