Adds an addition travel point to the bog to the Forest and Mountain for Vanderlin (#1880)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->
The tin, this adds one more travel point between the mountain and the
bog and the forest, it allows for better accessibility between areas.
The topography does not make sense, but Aberra is fine with it.


![PR1](https://github.com/user-attachments/assets/76a3b545-d4e3-4f9d-a641-ac92fe09e5a3)


![PR2](https://github.com/user-attachments/assets/df11e656-0d74-49d5-8181-bef7a4413644)


![PR3](https://github.com/user-attachments/assets/c86e8bd9-ed5b-4447-9e58-d845680ac158)


![PR4](https://github.com/user-attachments/assets/e5efac18-6cb6-49bd-8124-af4e82c1cbfa)


![PR5](https://github.com/user-attachments/assets/0f3ee0d9-5c98-4488-a8ae-04b4ec34cc8f)


![pr6](https://github.com/user-attachments/assets/d0cfcb84-3fdf-455b-8738-64d9c79e46d3)


I slightly touched up the bog, it is slated for a full rework as well.
The main change is that the bog fort no longer has fun loot for the loot
goblins. The inhumen camp entrance was moved more south.


## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

More movement = Good

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->