Major customizer quality control removals + tiefling horn customization. (#1872)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Attention Hollowkin: It is time for THE PURGE!!
There was not much QA when adding them, apparently, so I have removed
the ones which outright do not fit and commented out ones with glaring
sprite/clipping issues. Deal with it until they're fixed, or get used to
them being gone- I'm sorry, but that's my job.

Tieflings: Horn customization options! Default horn colour set to grey
in preparation for Borbop's implementation of list-based colour
selection akin to hair.
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: AberrantQuerist <crystalseeker04@gmail.com>