Heart crit requires rib fracture first or specific bclass (#1842)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Chest applies a normal artery wound unless the ribs are fractured or a
bow/gun/pickaxe (for soul come on pickaxes are awful) is used.

Make the neck artery fatal to CRITICAL_WEAKNESS since its as lethal as a
heart crit.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
You cannot miss the chest, having an extra step (fracture) to basically
immediately kill someone is probably a good thing and gives maces more
of a use/requires more input then just stab chest.

Ranged weapons could do with a bit of love, if a bullet is going THROUGH
your chest it should be able to hit your heart regardless.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->