Container Crafting System (#1803)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
moves pot recipes over to a new type of system:
Container Crafts
These work by just putting the items in the inventory, so no weird
slapcrafting.
This also allows for optional items for recipes so we can make tea with
1 milk 1 sugar, or oatmeal with mashed lemons

Since this is already a hellscape of a pr ports:
https://github.com/tgstation/tgstation/pull/69703
https://github.com/tgstation/tgstation/pull/69376
https://github.com/tgstation/tgstation/pull/71730

grabs sprites from:
https://github.com/GeneralPantsuIsBadAtCoding/Azure-Peak/pull/2186
redoes how the boss fight works
<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
More flavor to cooking, better code overall, easier to use by players
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->