Featured Statistics DLC: Half Drows, Smiths and Gluttons (#1898)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

This PR does 4 things:

- Adds Half Drows to the census along with a new corresponding row
- Adds 7 new featured statistics - Food, Smiths, Forged Items, Flaws,
Crops, Bleeders and Spells
- Minor fixes and changes for the older statistics, like properly
filtering beverages and changing top gourmets to just top eaters
- Very minor infuence tweaks

Also improves gods' champions formatting and fixes two minor issues with
the schism event.

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

More statistics, cleanup of the old stats, maintenance work.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->