Evil Wizard Song (#1849)
## About The Pull Request

Adds a percent chance shitpost cmode song to court magician and lich,
ook told me to do this.

## Why It's Good For The Game

Its really funny.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->