Tiefling blood Tea! (#1914)
## About The Pull Request
adds tiefling blood tea,

Created by drying sugar on a rack with 11oz of tiefling blood nearby...
then boiling the tiefling sugar in a pot.



![image](https://github.com/user-attachments/assets/885b92b8-3fcb-45db-ba86-1876ac23b611)


![image](https://github.com/user-attachments/assets/757e76fc-4a4d-4917-bccd-9fd23b0e1b3c)


## Why It's Good For The Game
Tiefling blood wine is such a delicacy.... but what if we had sweet
tiefling blood tea.... i like tea
## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->