(SOUND) New Bandit Track (#1949)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Adds the song 'Deadly Shadows' by T87-Sulfurhead to Hedge Knight and
Knave.

Moves the old Hedge Knight song into the unused folder.

## Why It's Good For The Game

Repurposed the thieves guild theme as the PR is currently in cryostasis
for the foreseeable future. Antagonists deserve baller themes.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->