Abyssor Drip (#1964)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Adds Abyssor acolyte robes, and will add Abyssor templar armor once the
sprites are finished. (Made by karlgoblington)

![image](https://github.com/user-attachments/assets/d7832f0b-ab02-4f47-954a-b1b104474b51)

## Why It's Good For The Game

More outfits for the church, its either you drip or drown, but why not
both?

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->