Fixes silver not working (#1956)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Ye silver was not working because of the || since the user always had
vampire but not werewolf and also the inverse, so silver was basically
useless all this time

- Fixes silver not working
- Makes so werewolf also recieve the debuff status
- Makes so psycrosses of silver and all their subtypes, being all the
the ten crosses carry the silver stat.
- Change abit transfix text to show who is the target if they roll more
than the vampire and carry a psycross and the usual if they got magic or
holy knowledge


Closes: #1953 
Vamp spawn trying to hold silver


![vamp_spawn](https://github.com/user-attachments/assets/7b20ee6e-5d17-41fb-af30-e8eab501e72c)

Vamp spawn being attacked with silver


![vampire_spawn_attacked](https://github.com/user-attachments/assets/b78a9860-0d3a-4dbd-8c22-dc7f86c7a03f)


Non ascended vamp lord trying to hold silver


![vamp_lord](https://github.com/user-attachments/assets/73d7e6d7-5277-49f7-9bfd-06fdddfc00ed)

Non ascended vamp lord being hit by silver

![vamp_lord_struke](https://github.com/user-attachments/assets/e6cdcfde-d9b1-4e9c-8af8-a7e33a01261d)

Ascended vamp lord holding silver

![ascended](https://github.com/user-attachments/assets/6eac0e4a-441a-451d-a09c-8635dfc74ab4)

Someone attacking an ascended vamp lord with silver  


![attacking_ascended_vamp_lord](https://github.com/user-attachments/assets/2b9789be-5bd8-4b25-aa93-ca22a1f55c13)


Person with psycross rolling more than the vampire example:


![above_blood_roll](https://github.com/user-attachments/assets/f831f7a8-1685-4bb2-8dc1-be38cab6f7e1)

Werewolf being hit by silver, get the usual stun, fire stuff and the
silver curse debuff for 1 min and 40 seconds. (in the code they also
can't eat people to heal if they got the status)


![werewolf_stat](https://github.com/user-attachments/assets/ff19c28d-b44a-4ebc-979e-32852efe56c9)


## Why It's Good For The Game

Fix

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->