Rosewood Marsh Update (#1931)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

![image](https://github.com/user-attachments/assets/06dbc7a2-584c-4fca-93e5-f88b91d43b38)
![StrongDMM-2025-05-16 04 32
43](https://github.com/user-attachments/assets/cc4bbd95-cafa-4ba6-b28f-2f8b72563a76)

![image](https://github.com/user-attachments/assets/819f8ec3-da7a-4a11-96f6-1ba940805171)

Last image cannot be uploaded via strongDMM screenshot due to size of
file.

This update finally adds the last OtherZ level of rosewood into the
game!
Features of the update are the following:
-New marsh map with unique dungeon designs and an overall story based on
Zizo cultists and what could happen if they were to take over a city.
New boss at the end which is currently a reskinned troll, the dragon is
a dracolich and is designed to work with undead in the zizo chapel.
-Additions to the mountains, forest, and rosewood city for travel tiles
to work. Shortcut can be achieved by heading to bog to mountains now!
-Minor fixes from feedback to rosewood city.

## Why It's Good For The Game
Adds more diversity to rosewood's map over all. The mountains still
don't work due to AI issues and the marsh is here to solve the
adventurer's need to explore dangerous locations to obtain treasure and
die horribly! With this update there is secrets to find, fun custom
items that have unique descriptions but don't have different stats, and
some fun dungeons to explore! WIENER RAID!

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [X] You tested this on a local server.
- [X] This code did not runtime during testing.
- [X] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: CheffieGithub <113442598+CheffieGithub@users.noreply.github.com>