Handcart Tweaks Part 2 (#1702)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Following after #1627 , this refactors the code and adds a `do_after`
delay to the sieving functionality. You could pick up a gigantic pile of
ore all at once before, but now there's a delay to make it less
immershun-breaking.

## Why It's Good For The Game

Potato wanted this change, but the PR got merged before I could get it
in. It doesn't make too much sense for this function to happen
instantly, though the delay can be tuned if half a second is too high.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->