Special Trait - Changes and Addition (#1875)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
Remove Combat-Ready trait from Thick skin, Combat-Ready make it so you
don't lose stamina while combat mode is active, replace it with Critical
resistance, making the special trait have a bit more sense.
Remove Cunning Linguist, it's basically polyglot but shittier with no
buff to it.
Add the other languages to Polyglots(Zybantine, Old Psy and Thieve's
Cant(it is a language)
Wild night now spawn you while being drugged and intoxicated.
Remove Godless. Useless, we already have godless as a faith to choose,
replace it with graggazo. (It doesn't do anything, just a
different-unique god.)
Add Magician Monarch as a Special trait from another codebase that
removed it
Add Skeleton appeareance as a Special trait(no infinite stam though,
y'all are insane if you think you can receive this from that.)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
More traits are always funny.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->