Makes Brigand not so Shit (#2000)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Adds the following to Brigand:
Iron Breastplate (You cannot parry with a flail. As a result, he gets
hit alot.)
Chain Coif (Cloth Coif is virtually useless)
Shields 3 (He is a SHIELD USER. Why was he missing this?)

## Why It's Good For The Game

Brigand is a role that cannot dodge, and is also optionally a flail user
- no matter what though, he's a shield user.

As somebody who cannot dodge, most attacks are redirected to the chest.

As somebody who cannot always parry (flail user) his armor gets
shredded. Due to this, the Brigand is completely awful to play.

This makes it so he is actually playable for once.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->