Port proximity monitor updates which includes a shit load of prs (#2030)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
The main thing here is that we port
https://github.com/tgstation/tgstation/pull/62755 porting this included
a few things.

most important being removal of Uncross() and CheckExit() and replacing
those behaviours with connect_loc signals.
(https://github.com/tgstation/tgstation/pull/58188) and related fixes.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
This might let us use CREATE AND DESTROY soon very good

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [ ] You tested this on a local server.
- [ ] This code did not runtime during testing.
- [ ] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->