KFC - By harpies, for harpies.  (#2034)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request
1. Increases harpy intelligence. Corvids are smart, this also will
encourage them to play more service roles than to be shitters.
2. Species-based limb butchering. Harpies are now bird meat, half orcs
are strange meat.


![kfc1](https://github.com/user-attachments/assets/adb53e16-35a4-466c-94b5-583f4ccefca5)

![kfc2](https://github.com/user-attachments/assets/8c46a238-da89-4f26-8342-66c667514cb4)

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

## Why It's Good For The Game
Negative int doesn't make too much sense for a supposedly uplifted
species, also more flavour (literally) for Graggarites or perhaps those
who are desperate.
<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: AberrantQuerist <crystalseeker04@gmail.com>