Bugfixes and tweaks: Family, wizard robes, large cogwheels, dendor sacrifices (#2039)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

This PR fixes and tweaks the following:

- Wizard robes are now craftable, previously they had the same recipe
name as a wizard hat, making them unreachable to crafters
- Fixes several issues when examining family members, for example some
relations were missing or fathers saw their brothers as their sons and
so forth
- List family verb will now show real names of the family members, no
more "Unknown - Father" and such
- FamilyUI now defaults to ON if you are part of any family, no more
mothers not realizing they have a son and such
- Fixes opendream warning regarding dragon's attack
- Fixes crafting issues with large cogwheels
- Always call OnCrafted for all crafted items
- Fixes several issues with dendor sacrifices to prevent duplicit
activations

Fixes #662
Fixes #958
Fixes #2038
Fixes #1792
Fixes #1362

## Why It's Good For The Game

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

Bugfixes.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->