Expand the Vanderlin basin Fort and the surrounding area (#2027)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

Makes the boring, empty, seemingly unfinished(??) Vanderlin basin fort
into a minor dungeon filled with rousmen.

![image](https://github.com/user-attachments/assets/09dba492-d8a9-4c9d-b2b7-cf72aa266b15)

![image](https://github.com/user-attachments/assets/cc1e90a0-587c-4b63-9c65-bb27ca3ebb8d)

![image](https://github.com/user-attachments/assets/03b230d8-fc2a-4e6e-8a56-c71146619a75)

![image](https://github.com/user-attachments/assets/c6de5585-0bab-4357-a8cb-a8c45d6a04ad)

Update the area around the fort with more foliage and random herbs
instead of fixed ones

![image](https://github.com/user-attachments/assets/5eaff5d2-1b20-4d66-af9b-7ac1d1401772)

![image](https://github.com/user-attachments/assets/c1fa2709-14fd-4d1f-a334-0182d009da34)

Expands the basin area around the fort so there is a dense treeline
behind the forest garrison. The main reason for this is so there is more
movement freedom (No more giant wooden walls in the basin).

![image](https://github.com/user-attachments/assets/bb169d85-0726-4715-8f3a-13cb5c5f98b3)

Adds a mapping helper for destroying structures.

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

More dungeons can't be bad, can be expanded on in future because there
is still so much space in the basin and below it to use.

The big walls that needed like 9 tiles to block off all access from the
main forest and bog exits were not interesting in the slightest, PATROL
THE WOODS.

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->