(SOUND) New towner combat track, gives music to most pilgrim / peasant jobs, deprecates combat_dendor.ogg (#2069)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Adds the track 'Knowledge and Pain' by T87-Sulfurhead to various
pilgrim/peasant jobs, assigns many of those aforementioned jobs a theme.

Also deprecates combat_dendor.ogg. The Forest Garrison theme is pretty
much the dendorite theme, godspeed dendor.ogg.

## Why It's Good For The Game

So many roles have no music assigned to them whatsoever. That's boring,
and it detracts from the fact that we need shitty rock-eating peasants
to fuel the economy.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->