Give Rogue Mage (Bandit) some TLC (#2052)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

## About The Pull Request

Gives the Rogue Wizard the following:

A Dark Hood to evilmax.
An Iron Quarterstaff.
Chalk.
Dark Spectacles in their bag.
Athletics 3 so they aren't immediately fucked in a fight.
Rare chance to play evil wizard music as their combat theme.
A normal goddamn belt for once.
Removes the generic mage outfit when they're old too.

## Why It's Good For The Game

Rogue Mage just needs a little bit more gas in the tank to not blow
absolute chunks. This doesn't make them stronger, just less horrible to
play.

## Pre-Merge Checklist
<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->