Adds the Mercenary Guild Master, "Gaffer" Job (#597)
<!-- Write **BELOW** The Headers and **ABOVE** The comments else it may
not be viewable. -->
<!-- Note: PRs including balance changes authored by anyone other than
maintainers and official devs will not be considered. -->

<!-- Describe The Pull Request. Please be sure every change is
documented or this can delay review or prevent the PR from being merged!
-->

Adds the Mercenary guild master and the 5th member of the old adventurer
party (as the ranger) ((approved by Aberra)), the "Gaffer"

![Gaffer](https://github.com/user-attachments/assets/06443c07-5b1b-4499-9478-9fedf85b30e2)

![profo2](https://github.com/user-attachments/assets/761832bf-d811-4786-a39e-7085dd7b277f)

-adds HAILER and HAILERBOARDS, a machine that acts as a global notice
board, needs the burdened trait

![profo3](https://github.com/user-attachments/assets/f918febd-97e8-42b2-9bc5-18efcb5ba1d5)
(below one is the HAILER)

![hailerboard](https://github.com/user-attachments/assets/b611d7c3-3b98-45ef-b898-0b3d2d6c95b5)
(hailerboard)

-adds some changes to the mercenaries' guild

![profo4](https://github.com/user-attachments/assets/9c117910-ccb9-41c0-beed-3db69795353f)

![profo5](https://github.com/user-attachments/assets/de43ccf4-741c-43d7-bdc2-c670194954f8)
(gaffer's room)

-adds the ring of burden, the Merc guild master's cursed ring (grants
the burdened trait needed for the machines)

![AAAAA](https://github.com/user-attachments/assets/edb221b0-7b49-49e3-b2b0-7d0783bb660a)

-adds 3 variant of "headhooks" storage items that go on the hip and
store heads from iron, bronze and royal(gold and silver) and adds the
storage components for the 2 (royal uses bronze)

![headhook](https://github.com/user-attachments/assets/043b52ca-c5df-42c6-a7e1-f4fc3f0087ef)
(royal not included because the devsprite was so awful I decided to make
it uncraftable)

-adds various new monster heads to sell to the HEADEATER

![buhghh](https://github.com/user-attachments/assets/6b46d0fd-3b63-45d1-9ba5-8b02c6387299)
(all of these are dev sprites)

-adds lame HEADEATER stand in until lore solidifies and bigger mapping
changes are allowed, used to sell items with headprice(heads, and no
player ones) (requires burdened)

![yaghghh](https://github.com/user-attachments/assets/7cb9bd22-a7b8-45c4-9649-4aee6d05d4e6)
(its the above one this time)

-adds the Ring's servant, a servant class for the Gaffer and the guild
![buh
huh](https://github.com/user-attachments/assets/54eb607c-cdaf-402d-9fc3-3c740eb3a28c)

-adds the subdued statue, the thing the ring of burden returns to when
it's dropped or destroyed.
![no
ring](https://github.com/user-attachments/assets/5fa6b59e-26a2-4bd8-8d37-b5465a764012)

![ring](https://github.com/user-attachments/assets/4b8a27c4-950e-4caa-8c6f-7a3e57e13dd0)

-adds Gaffer into the merchant's boat bell list
-adds Gaffer's keys
-adds Gaffers Keyring (consisting of 4 mercenary keys and one gaffer)
-adds the Gaffer start landmark
-adds the Gaffer's tutorial book "don't be a gaff, a guild master's
manual"
-adds "burdened" trait for the ring bearer, allows the use of HEADEATER
and HAILER as well as seeing the headprices of heads
-adds 2 new stress events (for gaffer selling his old trophies, and
inspecting certain items with burdened trait)
-adds headprice vars
-adds the Gaffer's useless vanity sword
-adds a see through cosmetic eyepatch and gives it to the gaffer, gives
the see through eyepatch a sewing recipe
-adds gaffercombat music (thank you mechadh!)
-adds a few head trophies from the Gaffer's past(not included)
-peppers and salts HAILERBOARDS around the town, the travel point from
the murderwoods to the town, the bog gate, and the anvil travel point
back to the town
-changes the merc token to be signed by the gaffer
-changes some monster drops to include the heads as perfect butchers
-changes goblin, orc, rousman etc heads to use headprice instead of
sellprice

<!-- Please add a short description of why you think these changes would
benefit the game. If you can't justify it in words, it might not be
worth adding. -->

The Gaffer adds a new RP opportunity to the server as the Guild Master
of the Mercenaries (and adventurers), being responsible for managing
quests; making sure given quests are completed on time, making sure the
quest loot is delivered to the customer, acting as a broker between
client and Merc/Adventurer, building character profiles for who is best
suited for which task, eliminating unreliable/untrustworthy folk from
taking advantage of the client, potentially acting as a bridge between
the smithies and merchants for Mercs and Adventurers so they can go out
and RP/have adventures while their orders are being fulfilled on their
behalf, preventing merchants from killing Adventurers main income by
being competitor for goblin (and in the future other monsters) heads.
Essentially, making sure the Merc/Adventurer experience is far more
smooth running and killing the dead air RP of endless wait times people
have for anything basic before they are allowed to interact with people
proper, or go on epic quests.

needs sprites

<!-- Don't bother filling these in while creating your Pull Request,
just click the checkboxes after the Pull Request is opened and you are
redirected to the page. -->
- [x] You tested this on a local server.
- [x] This code did not runtime during testing.
- [x] You documented all of your changes.
<!-- Neither the compiler nor workflow checks are perfect at detecting
runtimes and errors. It is important to test your code/feature/fix
locally. -->

---------

Co-authored-by: tonty <39193182+tontyGH@users.noreply.github.com>
Co-authored-by: grungussuss <96586172+Sadboysuss@users.noreply.github.com>